pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "visitor-map-backend"
version = "0.1.0"
description = "Lightweight visitor tracking backend and map generator for GitHub Pages"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "httpx>=0.25.0",
    "folium>=0.14.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff"
]

[tool.setuptools]
py-modules = ["main", "routes", "database", "services", "map_render"]
